Add GEFS support to rocoto XML generation (#1601)
To facilitate this, a few classes were refactored into abstract
classes with child classes for each of the three `$NET`/`$MODE`
combinations (gfs/cycled, gfs/forecast-only, and gefs). These child
classes then implement the details particular to their configuration.

GEFS is implemented as only forecast jobs right now: a fcst job
intended for the control and a metatask of efcs jobs intended for the
perturbed members. A couple of additional config files were needed in
order to parse the configs; these were copied from the gfs versions.
Beyond that, changes have not yet been made to update any configs for
GEFS. For instance, in contrast to GFS, even the control member of
GEFS will want `MEMDIR` defined (to 'mem000').

Fixes #1577