Add new python-based offline UPP (#1676)
The global-workflow runs offline UPP to generate master (and optionally flux) grib2 files for:
- analysis
- GOES simulated satellite imagery
- WAFS products

WAFS has been removed from the global-workflow in a previous PR.

The offline UPP and product generation in the current workflow is done in a single job.  This PR aims to partition the offline UPP from the product generation (for atmos grib2 files).

The changes in this PR are not yet activated in the workflow.

The offline post is now completely yaml driven for all the variants of the upp run conditions.

In a future PR, product generation jobs will be added and subsequently current post-processing jobs will be retired.